local/airflow_docker/requirements.txt
# kafka-python==1.4.6
# psycopg2-binary==2.7.7
matplotlib==3.0.3
# pyspark==2.4.4
# psycopg2
azure-core==1.9.0
azure-storage-blob==12.6.0
google-cloud-storage
docker==4.0.2
apache-airflow[postgres, google]
apache-airflow-backport-providers-microsoft-azure

